pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amplifier-ux-analyzer"
version = "0.2.0"
description = "AI-powered UX screenshot analyzer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opencv-python>=4.5.0",
    "numpy>=1.19.0",
    "scikit-learn>=0.24.0",
    "scikit-image>=0.18.0",
    "easyocr>=1.4.0",
    "pillow>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
amplifier-ux-analyzer = "amplifier_ux_analyzer.cli.main:main"
